Use generated withdrawal keys for generate-and-register
