+++ teku/src/main/java/tech/pegasys/teku/cli/deposit/DepositRegisterCommand.java
@@ -21,6 +21,8 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
+import org.apache.tuweni.bytes.Bytes48;
+import picocli.CommandLine;
 import picocli.CommandLine.ArgGroup;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Mixin;
@@ -33,7 +35,9 @@ import tech.pegasys.signers.bls.keystore.KeyStoreLoader;
 import tech.pegasys.signers.bls.keystore.KeyStoreValidationException;
 import tech.pegasys.signers.bls.keystore.model.KeyStoreData;
 import tech.pegasys.teku.bls.BLSKeyPair;
+import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.bls.BLSSecretKey;
+import tech.pegasys.teku.cli.options.WithdrawalPublicKeyOptions;
 import tech.pegasys.teku.util.cli.PicoCliVersionProvider;
 
 @Command(
@@ -56,6 +60,9 @@ public class DepositRegisterCommand implements Runnable {
   @Mixin private RegisterParams registerParams;
 
   @ArgGroup(exclusive = true, multiplicity = "1")
+  private WithdrawalPublicKeyOptions withdrawalKeyOptions;
+
+  @ArgGroup(exclusive = true, multiplicity = "1")
   private ValidatorKeyOptions validatorKeyOptions;
 
   @Mixin private VerboseOutputParam verboseOutputParam;
@@ -72,14 +79,38 @@ public class DepositRegisterCommand implements Runnable {
       final Function<String, String> envSupplier,
       final CommandSpec spec,
       final RegisterParams registerParams,
-      final ValidatorKeyOptions validatorKeyOptions,
-      final String withdrawalKey) {
+      final ValidatorKeyOptions validatorKeyOptions) {
     this.shutdownFunction = shutdownFunction;
     this.envSupplier = envSupplier;
     this.spec = spec;
     this.registerParams = registerParams;
     this.validatorKeyOptions = validatorKeyOptions;
     this.verboseOutputParam = new VerboseOutputParam(true);
+  }
+
+  public BLSPublicKey getWithdrawalPublicKey() {
+    if (withdrawalKeyOptions.withdrawalKey != null) {
+      return BLSPublicKey.fromBytesCompressed(
+          Bytes48.fromHexString(withdrawalKeyOptions.withdrawalKey));
+    } else if (withdrawalKeyOptions.withdrawalKeystoreFile != null) {
+      return getWithdrawalKeyFromKeystore();
+    } else {
+      // not meant to happen
+      throw new IllegalStateException("Withdrawal Key Options are not initialized");
+    }
+  }
+
+  private BLSPublicKey getWithdrawalKeyFromKeystore() {
+    try {
+      final KeyStoreData keyStoreData =
+          KeyStoreLoader.loadFromFile(withdrawalKeyOptions.withdrawalKeystoreFile.toPath());
+      return BLSPublicKey.fromBytesCompressed(Bytes48.wrap(keyStoreData.getPubkey()));
+    } catch (final KeyStoreValidationException e) {
+      throw new CommandLine.ParameterException(
+          spec.commandLine(),
+          "Error: Unable to get withdrawal key from keystore: " + e.getMessage(),
+          e);
+    }
   }
 
   @Override
@@ -89,7 +120,7 @@ public class DepositRegisterCommand implements Runnable {
     try (final RegisterAction registerAction =
         registerParams.createRegisterAction(verboseOutputParam.isVerboseOutputEnabled())) {
       registerAction.displayConfirmation(1);
-      registerAction.sendDeposit(validatorKey).get();
+      registerAction.sendDeposit(validatorKey, getWithdrawalPublicKey()).get();
     } catch (final Throwable t) {
       SUB_COMMAND_LOG.sendDepositFailure(t);
       shutdownFunction.accept(1);
